transformation to int type finished

+++ includes/fdf.h
@@ -6,7 +6,7 @@
 /*   By: thakala <thakala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/24 11:09:46 by thakala           #+#    #+#             */
-/*   Updated: 2022/03/31 09:28:21 by thakala          ###   ########.fr       */
+/*   Updated: 2022/03/31 09:31:28 by thakala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,16 +41,16 @@ typedef struct s_point
 
 typedef struct s_line
 {
-	int			*line;
-	uint64_t	point_count;
-	uint64_t	size;
+	int		*line;
+	int		point_count;
+	int		size;
 }	t_line;
 
 typedef struct s_map
 {
 	t_line		*lines;
-	uint64_t	line_count;
-	uint64_t	size;
+	int			line_count;
+	int			size;
 	// int64_t	*line_lengths;
 }	t_fdf_map;
